pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.4",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]
